+++ packages/gamut-styles/src/index.ts
@@ -1,13 +1,19 @@
 import '@emotion/react';
 
-import { Theme as GamutTheme } from './theme';
+import { theme } from './theme';
+
+export * from './GamutProvider';
 
 export * from './cache';
 export * from './variables';
+export * from './styles';
+export * from './globals';
 export * from './utilities';
-export * from './system';
-export * from './theme';
+export * from './props';
+export { theme } from './theme';
+
+export type ThemeShape = typeof theme;
 
 declare module '@emotion/react' {
-  export interface Theme extends GamutTheme {}
+  export interface Theme extends ThemeShape {}
 }