+++ packages/gamut/src/Form/Select.tsx
@@ -1,24 +1,78 @@
-import cx from 'classnames';
+import {
+  ArrowChevronDownIcon,
+  MiniChevronDownIcon,
+} from '@codecademy/gamut-icons';
+import { variant } from '@codecademy/gamut-styles';
+import { css } from '@emotion/react';
+import styled from '@emotion/styled';
 import { each, isArray, isObject } from 'lodash';
-import React, { forwardRef, ReactNode, SelectHTMLAttributes } from 'react';
+import React, {
+  ChangeEvent,
+  forwardRef,
+  ReactNode,
+  SelectHTMLAttributes,
+  useState,
+} from 'react';
 
-import styles from './styles/Select.module.scss';
+import { Box, FlexBox } from '../Box';
+import { conditionalStyles, formFieldStyles } from './styles/shared';
 
-export type SelectProps = SelectHTMLAttributes<HTMLSelectElement> & {
+export type SelectWrapperProps = SelectHTMLAttributes<HTMLSelectElement> & {
   error?: boolean;
   htmlFor?: string;
   options?: string[] | Record<string, number | string>;
   id?: string;
+  sizeVariant?: 'small' | 'base';
 };
 
-export const Select = forwardRef<HTMLSelectElement, SelectProps>(
-  (props, ref) => {
-    const className = cx(
-      styles.Select,
-      props.className,
-      props.error && styles.error
-    );
-    const { options, error, id, ...rest } = props;
+export interface SelectProps extends SelectWrapperProps {
+  activated?: boolean;
+}
+
+const selectSizeVariants = variant({
+  default: 'base',
+  prop: 'sizeVariant',
+  variants: {
+    small: {
+      height: '2rem',
+      paddingX: 8,
+      paddingY: 0,
+    },
+    base: {
+      height: 'auto',
+    },
+  },
+});
+
+const SelectBase = styled.select<SelectProps>`
+  ${formFieldStyles}
+  ${conditionalStyles}
+  ${selectSizeVariants}
+  cursor: pointer;
+  display: block;
+  -moz-appearance: none;
+  -webkit-appearance: none;
+  appearance: none;
+`;
+
+const selectIconStyles = css`
+  pointer-events: none;
+`;
+
+const SelectIcon = styled(ArrowChevronDownIcon)(selectIconStyles);
+const MiniSelectIcon = styled(MiniChevronDownIcon)(selectIconStyles);
+
+export const Select = forwardRef<HTMLSelectElement, SelectWrapperProps>(
+  (
+    { className, defaultValue, options, error, id, sizeVariant, ...rest },
+    ref
+  ) => {
+    const [activated, setActivated] = useState(false);
+
+    const changeHandler = (event: ChangeEvent<HTMLSelectElement>) => {
+      rest?.onChange?.(event);
+      setActivated(true);
+    };
 
     let selectOptions: ReactNode[] = [];
 
@@ -43,23 +97,40 @@ export const Select = forwardRef<HTMLSelectElement, SelectProps>(
     }
 
     return (
-      <div className={className}>
-        <svg className={styles.selectIcon}>
-          <path
-            d="M1.175 0L5 3.825 8.825 0 10 1.183l-5 5-5-5z"
-            fill="#3E3E40"
-          />
-        </svg>
-        <select
+      <Box
+        position="relative"
+        width="100%"
+        textColor={error ? 'red' : 'navy'}
+        minWidth="7rem"
+        className={className}
+      >
+        <FlexBox
+          paddingRight={12}
+          alignItems="center"
+          position="absolute"
+          right="0"
+          top="0"
+          bottom="0"
+        >
+          {sizeVariant === 'small' ? (
+            <MiniSelectIcon size={12} />
+          ) : (
+            <SelectIcon size={16} />
+          )}
+        </FlexBox>
+        <SelectBase
           {...rest}
-          className={styles.selectInput}
-          defaultValue={props.defaultValue || ''}
-          id={id || props.htmlFor}
+          defaultValue={defaultValue || ''}
+          id={id || rest.htmlFor}
           ref={ref}
+          error={error}
+          sizeVariant={sizeVariant}
+          activated={activated}
+          onChange={(event) => changeHandler(event)}
         >
           {selectOptions}
-        </select>
-      </div>
+        </SelectBase>
+      </Box>
     );
   }
 );