+++ packages/gamut/src/Markdown/libs/overrides/MarkdownAnchor/index.tsx
@@ -1,9 +1,9 @@
 /* eslint-disable jsx-a11y/anchor-has-content */
-import React, { HTMLAttributes } from 'react';
+import React, { ComponentProps } from 'react';
 
 import { Anchor } from '../../../../Anchor';
 
-export interface MarkdownAnchorProps extends HTMLAttributes<HTMLAnchorElement> {
+export interface MarkdownAnchorProps extends ComponentProps<typeof Anchor> {
   href: string;
 }