+++ packages/styleguide/.storybook/addons/system/propMeta.ts
@@ -1,6 +1,6 @@
-import * as system from '@codecademy/gamut-styles/src/system';
+import * as system from '@codecademy/gamut-styles/src/props';
 
-const { shouldForwardProp, properties, variant, ...groups } = system;
+const { properties, variant, ...groups } = system;
 
 export const PROP_GROUPS = groups;