+++ packages/styleguide/.storybook/decorators/theme.tsx
@@ -1,9 +1,6 @@
 import React from 'react';
-import { CacheProvider, ThemeContext } from '@emotion/react';
 
-import { theme, createEmotionCache } from '@codecademy/gamut-styles';
-
-const cache = createEmotionCache();
+import { GamutProvider } from '@codecademy/gamut-styles/src';
 
 /**
  * Story functions must be called as a regular function to avoid full-remounts
@@ -11,11 +8,15 @@ const cache = createEmotionCache();
  */
 
 export const withEmotion = (Story: any) => {
-  return process.env.NODE_ENV === 'test' ? (
-    <ThemeContext.Provider value={theme}>{Story()}</ThemeContext.Provider>
-  ) : (
-    <CacheProvider value={cache}>
-      <ThemeContext.Provider value={theme}>{Story()}</ThemeContext.Provider>
-    </CacheProvider>
-  );
+  // Always give iframes the full provider
+  if (process.env.NODE_ENV === 'test') {
+    return (
+      <GamutProvider useCache={false} useGlobals={false}>
+        {Story()}
+      </GamutProvider>
+    );
+  }
+
+  // Wrap all stories in minimal provider
+  return <GamutProvider>{Story()}</GamutProvider>;
 };