+++ packages/gamut-styles/CHANGELOG.md
@@ -3,6 +3,78 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [9.1.0](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-styles@9.0.1...@codecademy/gamut-styles@9.1.0) (2021-03-23)
+
+
+### Features
+
+* **GamutProvider:** Adds single provider for Theme + Cache + Globals + Assets [WEB-1352] [GM-173] ([9eac700](https://github.com/Codecademy/client-modules/commit/9eac7009b4d773621a0993cc125d6878b76977ce))
+
+
+
+### [9.0.1](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-styles@9.0.0...@codecademy/gamut-styles@9.0.1) (2021-03-19)
+
+**Note:** Version bump only for package @codecademy/gamut-styles
+
+
+
+
+
+## [9.0.0](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-styles@8.3.2...@codecademy/gamut-styles@9.0.0) (2021-03-18)
+
+
+### ⚠ BREAKING CHANGES
+
+* **ContentContainer:** update responsive padding
+
+### Features
+
+* **ContentContainer:** update responsive padding ([4f66f22](https://github.com/Codecademy/client-modules/commit/4f66f22a3572f1e721b5da0a8c41ebe3dd385279))
+
+
+
+### [8.3.2](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-styles@8.3.1...@codecademy/gamut-styles@8.3.2) (2021-03-16)
+
+**Note:** Version bump only for package @codecademy/gamut-styles
+
+
+
+
+
+### [8.3.1](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-styles@8.3.0...@codecademy/gamut-styles@8.3.1) (2021-03-04)
+
+
+### Bug Fixes
+
+* **styles:** Run emotion plugin ([c79b818](https://github.com/Codecademy/client-modules/commit/c79b818865a6ce370393543863ccbdf3cb46f5b7))
+
+
+
+## [8.3.0](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-styles@8.2.2...@codecademy/gamut-styles@8.3.0) (2021-03-04)
+
+
+### Features
+
+* **ThemeProvider:**  Adds GamutThemeProvider component and initial Global styles ([f4fd13c](https://github.com/Codecademy/client-modules/commit/f4fd13ce04762f75517117372c76d189ee0f0b99))
+
+
+
+### [8.2.2](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-styles@8.2.1...@codecademy/gamut-styles@8.2.2) (2021-03-04)
+
+**Note:** Version bump only for package @codecademy/gamut-styles
+
+
+
+
+
+### [8.2.1](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-styles@8.2.0...@codecademy/gamut-styles@8.2.1) (2021-03-04)
+
+**Note:** Version bump only for package @codecademy/gamut-styles
+
+
+
+
+
 ## [8.2.0](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-styles@8.1.1...@codecademy/gamut-styles@8.2.0) (2021-03-02)